pyproject.toml
[project]
name = "baat-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.3",
    "chromadb>=1.5.2",
    "deepgram-sdk>=6.0.1",
    "elevenlabs>=2.38.1",
    "google-cloud-speech>=2.37.0",
    "google-cloud-texttospeech>=2.34.0",
    "langchain>=1.2.10",
    "langchain-anthropic>=1.3.4",
    "langchain-classic>=1.0.1",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.10",
    "langgraph>=1.0.10",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.2.3",
    "webrtcvad-wheels>=2.0.14",
    "websockets>=16.0",
]
